Move VUMI_INSTALLED_APPS and VUMI_INSTALLED_ROUTERS into go.config helper functions.
