+++ go/base/tests/test_utils.py
@@ -2,33 +2,13 @@
 
 from go.base.tests.utils import VumiGoDjangoTestCase
 from go.errors import UnknownConversationType, UnknownRouterType
+
+import go.base.utils
 from go.base.utils import (
-    get_conversation_pkg, get_conversation_definition,
-    get_conversation_view_definition,
-    get_router_pkg, get_router_definition,
-    get_router_view_definition)
+    get_conversation_view_definition, get_router_view_definition)
 
 
 class ConversationDefinitionHelpersTestCase(VumiGoDjangoTestCase):
-    def test_get_conversation_pkg(self):
-        pkg = get_conversation_pkg('bulk_message', ['definition'])
-        self.assertEqual(pkg.__name__, 'go.apps.bulk_message')
-
-    def test_get_conversation_pkg_fails(self):
-        self.assertRaises(UnknownConversationType,
-                          get_conversation_pkg, 'unknown', ['definition'])
-
-    def test_get_conversation_definition(self):
-        conv_def = get_conversation_definition('bulk_message')
-        from go.apps.bulk_message.definition import ConversationDefinition
-        self.assertTrue(isinstance(conv_def, ConversationDefinition))
-        self.assertEqual(conv_def.conversation_type, 'bulk_message')
-
-    def test_get_conversation_definition_with_conv(self):
-        dummy_conv = object()
-        conv_def = get_conversation_definition('bulk_message', dummy_conv)
-        self.assertTrue(conv_def.conv is dummy_conv)
-
     def test_get_conversation_view_definition(self):
         view_def = get_conversation_view_definition('bulk_message')
         from go.apps.bulk_message.view_definition import (
@@ -46,30 +26,13 @@ class ConversationDefinitionHelpersTestCase(VumiGoDjangoTestCase):
             get_conversation_view_definition, 'not_droids')
 
     def test_get_conversation_view_definition_obsolete_conv_type(self):
-        view_def = get_conversation_view_definition('wikipedia_sms')
-        self.assertEqual(view_def._conv_def.conversation_type, 'wikipedia_sms')
+        self.monkey_patch(go.base.utils, 'obsolete_conversation_types',
+                   lambda: set(['old_conv']))
+        view_def = get_conversation_view_definition('old_conv')
+        self.assertEqual(view_def._conv_def.conversation_type, 'old_conv')
 
 
 class RouterDefinitionHelpersTestCase(VumiGoDjangoTestCase):
-    def test_get_router_pkg(self):
-        pkg = get_router_pkg('keyword', ['definition'])
-        self.assertEqual(pkg.__name__, 'go.routers.keyword')
-
-    def test_get_router_pkg_fails(self):
-        self.assertRaises(UnknownRouterType,
-                          get_router_pkg, 'unknown', ['definition'])
-
-    def test_get_router_definition(self):
-        router_def = get_router_definition('keyword')
-        from go.routers.keyword.definition import RouterDefinition
-        self.assertTrue(isinstance(router_def, RouterDefinition))
-        self.assertEqual(router_def.router_type, 'keyword')
-
-    def test_get_router_definition_with_router(self):
-        dummy_router = object()
-        router_def = get_router_definition('keyword', dummy_router)
-        self.assertTrue(router_def.router is dummy_router)
-
     def test_get_router_view_definition(self):
         view_def = get_router_view_definition('keyword')
         from go.routers.keyword.view_definition import (
@@ -80,3 +43,14 @@ class RouterDefinitionHelpersTestCase(VumiGoDjangoTestCase):
         dummy_router = object()
         view_def = get_router_view_definition('keyword', dummy_router)
         self.assertTrue(view_def._router_def.router is dummy_router)
+
+    def test_get_router_view_definition_unknown_conv_type(self):
+        self.assertRaises(
+            UnknownRouterType,
+            get_router_view_definition, 'not_the_router_we_are_looking_for')
+
+    def test_get_router_view_definition_obsolete_router_type(self):
+        self.monkey_patch(go.base.utils, 'obsolete_router_types',
+                   lambda: set(['old_router']))
+        view_def = get_router_view_definition('old_router')
+        self.assertEqual(view_def._router_def.router_type, 'old_router')
